Update spring social connectors.
- Fix bug with missing "filterDao" (database.xml)
- Updates spring social connectors to latest releases.
- Facebook connector: updated scopes to current API (see ProfileController.java)
- Potential issue: Facebook social controller no longer receives username from auth provider.
This may affect new users. As a fallback solution, if username is null but email is not
null, the name in the email is assumed to be the username. This should not affect existing
Facebook users, as nquireit identifies these by provider user id, which in the case of
Facebook is a number not related to the email or username.